+++ PRISMDatabaseUtils/PostgresSQL/PostgresDBTools.cs
@@ -298,7 +298,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryScalar(
             string sqlQuery,
             out object queryResult,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             int timeoutSeconds = -1,
             [CallerMemberName] string callingFunction = "UnknownMethod")
@@ -328,7 +328,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryScalar(
             DbCommand cmd,
             out object queryResult,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             [CallerMemberName] string callingFunction = "UnknownMethod")
         {
@@ -432,7 +432,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResults(
             string sqlQuery,
             out List<List<string>> lstResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int maxRowsToReturn = 0,
             int retryDelaySeconds = 5,
             int timeoutSeconds = -1,
@@ -466,7 +466,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResults(
             DbCommand cmd,
             out List<List<string>> lstResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int maxRowsToReturn = 0,
             int retryDelaySeconds = 5,
             [CallerMemberName] string callingFunction = "UnknownMethod")
@@ -525,7 +525,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResultsDataTable(
             string sqlQuery,
             out DataTable queryResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             int timeoutSeconds = -1,
             [CallerMemberName] string callingFunction = "UnknownMethod")
@@ -555,7 +555,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResultsDataTable(
             DbCommand cmd,
             out DataTable queryResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             [CallerMemberName] string callingFunction = "UnknownMethod")
         {
@@ -589,7 +589,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResultsDataSet(
             string sqlQuery,
             out DataSet queryResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             int timeoutSeconds = -1,
             [CallerMemberName] string callingFunction = "UnknownMethod")
@@ -619,7 +619,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public bool GetQueryResultsDataSet(
             DbCommand cmd,
             out DataSet queryResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5,
             [CallerMemberName] string callingFunction = "UnknownMethod")
         {
@@ -652,7 +652,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         private bool GetQueryResults(
             DbCommand cmd,
             Action<NpgsqlCommand> readMethod,
-            short retryCount,
+            int retryCount,
             int retryDelaySeconds,
             string callingFunction)
         {
@@ -746,7 +746,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         private int ExecuteSPData(
             DbCommand spCmd,
             Action<NpgsqlCommand> readMethod,
-            short retryCount,
+            int retryCount,
             int retryDelaySeconds)
         {
             if (!(spCmd is NpgsqlCommand sqlCmd))
@@ -889,7 +889,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public int ExecuteSPData(
             DbCommand spCmd,
             out List<List<string>> lstResults,
-            short retryCount = 3,
+            int retryCount = 3,
             int maxRowsToReturn = 0,
             int retryDelaySeconds = 5)
         {
@@ -942,7 +942,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public int ExecuteSPDataTable(
             DbCommand spCmd,
             out DataTable results,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5)
         {
             var queryResults = new DataTable();
@@ -970,7 +970,7 @@ namespace PRISMDatabaseUtils.PostgresSQL
         public int ExecuteSPDataSet(
             DbCommand spCmd,
             out DataSet results,
-            short retryCount = 3,
+            int retryCount = 3,
             int retryDelaySeconds = 5)
         {
             var queryResults = new DataSet();
